Area ID in output info, some docs & cleanup
