+++ lib/round_spot.dart
@@ -1,9 +1,9 @@
 library round_spot;
 
 export 'src/entry_point.dart';
-export 'src/models/config.dart';
-export 'src/models/heat_map_style.dart';
+export 'src/models/config/config.dart';
+export 'src/models/config/heat_map_style.dart';
+export 'src/models/config/output_type.dart';
 export 'src/models/output_info.dart';
-export 'src/models/output_type.dart';
 export 'src/navigator_observer.dart';
 export 'src/widgets/detector.dart';
