+++ lib/src/components/processors/numerical_processor.dart
@@ -1,5 +1,5 @@
 import '../../models/session.dart';
-import '../../utils/export_utils.dart';
+import '../../utils/utils.dart';
 
 import 'session_processor.dart';
