+++ lib/src/entry_point.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/widgets.dart';
 
-import 'models/config.dart';
+import 'models/config/config.dart';
 import 'models/output_info.dart';
 import 'utils/components.dart';
 import 'widgets/detector.dart';
@@ -11,13 +11,14 @@ import 'widgets/lifecycle_observer.dart';
 typedef HeatMapCallback = void Function(Uint8List data, OutputInfo info);
 typedef NumericCallback = void Function(String data);
 
-Widget initialize(
-    {required Widget child,
-    Config? config,
-    HeatMapCallback? heatMapCallback,
-    NumericCallback? numericCallback}) {
+Widget initialize({
+  required Widget child,
+  Config? config,
+  HeatMapCallback? heatMapCallback,
+  NumericCallback? numericCallback,
+}) {
   initializeComponents(config, heatMapCallback, numericCallback);
-  return LifecycleObserver(child: Detector(child: child, areaID: 'global'));
+  return LifecycleObserver(child: Detector(child: child, areaID: ''));
 }
 
 Config get config => S.get<Config>();
