+++ lib/src/models/config/output_type.dart
