+++ lib/src/models/session.dart
@@ -1,21 +1,25 @@
 import 'dart:math';
 import 'dart:ui';
 
+import '../utils/utils.dart';
 import 'event.dart';
 import 'output_info.dart';
 
+/// Represents a record of user interactions
+/// with a particular [area] on some [page].
 class Session implements OutputInfo {
-  final List<Event> _events = [];
-  final String name;
+  final String page;
+  final String area;
   final int startTime;
   int endTime;
-  Image? screenSnap;
 
+  Image? screenSnap;
+  final List<Event> _events = [];
   List<Event> get events => _events;
 
-  Session({required this.name})
-      : startTime = DateTime.now().millisecondsSinceEpoch,
-        endTime = DateTime.now().millisecondsSinceEpoch;
+  Session({required this.page, required this.area})
+      : startTime = getTimestamp(),
+        endTime = getTimestamp();
 
   void addEvent(Event event) {
     _events.add(event);
@@ -23,8 +27,9 @@ class Session implements OutputInfo {
   }
 
   Map<String, dynamic> toJson() => {
-        'name': name,
+        'page': page,
+        'area': area,
         'span': {'start': startTime, 'end': endTime},
-        'events': [for (var event in _events) event.toJson()]
+        'events': [for (var event in _events) event.toJson()],
       };
 }
